tests/sources/ostree.py: Skip GPG test on platforms which dont support our key
We use some pre-generated gpg homedir data which we created on another
platform and it is unclear whether the homedir is compatible across hosts
and gpg versions.

Our key is not recognized on centos 7, so we now just add a check and
skip the test in the case that the key is not recognized.