tests/sources/git.py: Workaround ancient git on centos 7
Back in the days, git used to work very subtly differently than it does
in these days... and the test for invalid-submodule warnings is reporting
the inconsistent-submodule warnings instead.

This is not significantly important enough to address in the bst-1 branch
so we will just disable this invalid-submodule test in bst-1.