.github/compose/ci.docker-compose.yml: Wholesale disable test parallelism
As of recently, the test_deterministic_source_umask test from
src/buildstream/testing/_sourcetests/source_determinism.py has been hanging.

This is spuriously reproducible if you run the containers locally on your
linux machine and try to run tox repeatedly with the following invocation:

  tox -e py37-plugins -- -n 4 -k "test_deterministic_source_umask"

However, tests are hanging consistently in github CI, often with
the last visible line readable in CI being the running of these source
tests.

It appears that the only way to have tests pass at all in github now is
to disable the pytest-xdist parallelism.