_workspaces.py: Use bst2 specific data directory for private project state
This ensures that bst-1 and bst-2 do not step on eachothers toes when working
on the same directory, which can typically happen when a project is in
transition to bst-2, and a user needs to work on multiple branches.

When transitioning to bst2, workspaces which appear to be open in bst1 are
ignored by bst2, and workspaces opened with bst2 are ignored by bst1, however
both can be configured to use the same directory for a given workspace if
the user requires this.

This also updates the test case: tests/frontend/workspace.py

Fixes #1266