_scheduler/queues/queue.py: Prioritize enqueue order
The queue processing order was long since modified to ignore the
insertion order and use the depth assigned with Element._set_depth().

This patch adds the insertion order as the second criteria for sorting
elements (after depth), rather than falling back on processing elements
alphabetically.

This changes the behavior of non-build sessions such that elements
are processed in an order similar to the order in which they are listed
in the pipeline summaries.

This consequently also allows us to reuse some of the expected ordering
in some test cases which were using `--deps plan`, which is being removed.