Strict element naming format.
This patch removes the fatal(able) warnings for bad element names,
and instead makes BuildStream strict about element naming all around
such that we can now rely on elements being suffixed with ".bst" and
not contain invalid characters that are unsupported as filenames on some
platforms.

Summary of changes:

  * types.py: Removed associated CoreWarning entries

  * exceptions.py: Added corresponding LoadErrorReason entries

  * _loader/loader.py: Raise an error instead of a configurable warning
                       upon encountering invalid element names.

We also now provide the provenance of the reference to the
invalid element name in error messages (if it is not toplevel).