fuse: Pass through the underlying filesystem's inode numbers
This makes stat(2) and similar system calls return the underlying
filesystem's inode number instead of dynamically generating them.

This is a requirement of the main purpose of this change: returning
a consistent inode number in directory entries returned by
readdir(2).

Before this change, libfuse would always return the value 0xffffffff
(FUSE_UNKNOWN_INO) in the d_ino field for each directory entry
returned by readdir(2).

Unfortunately, we need to add a call to os.stat() in order to get the
inode number in Python, because the output of os.listdir() doesn't
expose this value, and os.scandir() can't be used as an alternative
since it unconditionally skips the '.' and '..' entries.