node.pyx: Fix node composition to not leak mutable data onto targets
As we've observed in #1360, it occurs that in composition, instead of
overwriting the target nodes with copies of source values, we insert the
mutable source values directly onto the target, this leaves the source
nodes vulnerable to future mutations when the code goes on to mutate
the resulting composited target.

Fixes #1360.