tests/frontend/completions.py: Make artifact completion test stable
The order in which artifacts are reported in the bash completion list
is not guaranteed to be a stable order (and nothing really requires this
to be), so update the test to allow both possible orders.

This fixes the test randomly failing where artifacts were not being
reported in the expected order.