storage/directory.py: Removed mark_unmodified/list_modified_paths
This was much implementation plumbing just for the sake of showing
which files were modified due to integration commands being run in
the compose plugin.

I think we're better off not carrying this weight in the new API
unless there is a strong motivation to bring it back in the future.

This removes a test case which was testing this functionality, along
with the implementations in the cas/file directory backends.