storage: Raise only DirectoryError from Directory backends
The API advertizes that it will raise DirectoryError for system errors,
better to raise DirectoryError than FileNotFoundError or other native python
system errors.

With the exception that we still raise ValueError for clear cut programming
errors which should not be handled by the caller but rather their code should
be fixed.