Remove duplicate code backing `bst artifact list-contents`
There was a whole separate code path to poke into the private details
of CasBasedDirectory for the purpose of listing artifact contents.

Instead, just use the Directory API directly to implement this now
and ditch all the extra redundant codepaths.

Interestingly, this corrects the tests/frontend/artifact_list_contents.py
test which was expecting a file size of 107 bytes, which is actually
only 28 bytes long.