storage/directory.py: Remove `can_link` parameter from Directory.import_files()
This is a historic artifact from when we used to hardlink files into a staging
area which is mostly rendered obsolete by allowing buildbox to handle these
details, it is ignored by the CasBasedDirectory completely, and we don't use
it anywhere with the FileBasedDirectory.

Remove the option entirely from public and private import_files() codepaths.