tests/frontend/buildcheckout.py: Remove obsolete test case
This test dates back to 2017 in commit 10847a66dc0748696829d5fd18ad17043c2c6878
and checks for there to be an error when depending on an artifact which had
installed files where the build directory is presumed to be.

Upon investigation, the test detects the expected error but for the wrong
reason, there is no conflict when staging files to the build directory but
instead an error is reported because there is no /install-root in the sandbox
when the manual build element completes it's build.

Remove this.