_cas/casdprocessmanager.py: Assert minimal required version of buildbox-casd
Here we pass along the Messenger object to the casdprocessmanager so that
we can issue a warning on startup in the case that we are unable to properly
determine the installed buildbox-casd version.

An error is raised if we successfully identify the installed version and
it is less than the required version.

Fixes #1189

Note this also updates the internal cascache test which itself generates
a dummy shell script to impersonate buildbox-casd, this change causes
that generated script to support the `--version` argument with a large
enough version (2.0.0).