Clean the build directory after a successful build
Capturing the build directory takes time and uses disk space of the CAS
cache. If the `cache-buildtrees` option is set to `auto` (the default)
or `never`, the contents of the build directory can be removed after a
successful build as the build tree is no longer needed, avoiding the
capture costs.