element.py: Retry non-strict failed builds where appropriate
In the case that we are building in non-strict mode and encounter a failed
build artifact, we discard the build artifact if the loaded artifact's
cache key differs from the cache key we would have otherwise assigned it.

Effectively this causes non-strict failed artifacts to be retried in
non-strict mode whenever the dependencies change.

Fixes #1460