_context.py, _stream.py: Add global "build" flag.
For a long time we have been getting on well without this, however in this
branch is becomes apparent that we require some state to inform conditional
logic about whether the invocation is concerned with building or not.

This state could evolve into a "mode" or such in the future if the need
arises, but for now we're only concerned with whether we are building.

Specifically in relation to #1460, we want to discard and ignore cached
failed build artifacts and retry building them in the case of `bst build`,
but we don't want to discard that artifact when shelling into the failed
build with `bst shell --build` (or in any other situations like showing
the failed state with `bst show` or sharing the failed artifact with
`bst artifact push` etc).