element.py: Lower expectations when discarding failed builds in non-strict mode
We were trying to compare the potential strong key against the artifact
strong key, but the potential strong key is not guaranteed to be known when
loading the artifact.

For now, we just retry failed builds in non-strict mode as long as the failed
build artifact key is not equal to the element's strict key, as we know that
dependencies *might* have changed since the failed build occurred.