node.pyx, _yaml.pyx: Support setting None values on MappingNode objects
This patch adds support for:
  * Setting none values on Node structures, e.g. foo["bar"] = None
  * Saving Node structures which contain None values to YAML

When saving None values to YAML, we save them as empty strings. This is
a valid None value by YAML specifications, but will later be loaded by
BuildStream as an empty string.

This change is required in order to address #1685, which requires the
ability to set None values in MappingNode objects.