Fix side effects of loading YAML booleans as strings
In the previous YAML behavior, we would serialize booleans as strings but
load booleans as boolean values, but after porting to the new ruamel API
it appears difficult to achieve this behavior in advance of a new ruamel
release including the fix https://sourceforge.net/p/ruamel-yaml/tickets/341/

This patch ensures buildstream at least internally uses the _yaml APIs
to access booleans loaded from the yaml so that tests pass.

Plugins should in theory be using Plugin.node_get_member() and such APIs
to access booleans already, but if they do not, they will break with this
branch.