_frontend/app.py: Fixed missing messages in frontend main logging
The threaded scheduler causes the frontend message handling to be
multi-threaded, as messages are no longer passed through a message queue
but directly propagated to the frondend.

This, in conjunction with message throttling means that we need to use
a lock to protect the cached message, otherwise sometimes a message gets
cached at the same time we finish printing a message and clear the cache.

This patch also ensures that we flush the message cache more consistently
when exiting the scheduler, and causes local app.py messages to be flushed
immediately just in case.