element.py, _frontend/widget.py: Use _yaml.roundtrip_dump_string()
Instead of deprecated round_trip_dump() from ruamel.yaml.

This will silence a lot of the remaining pesky deprecation warnings after
running tests in tox.

This changes the output of `bst show --format '%{deps}'`, such that
the YAML output is formatted as:

  - dep-a.bst
  - dep-b.bst

Instead of [dep-a.bst, dep-b.bst], which are both valid YAML so this is
not technically an API break, however the tests/frontend/show.py test
needed to be updated.