Fail early if `buildbox-run` is present but non-functional
If buildbox-run is missing, BuildStream falls back to "dummy" sandbox
which allows the program to start, and causes build command execution
to fail with a "missing sandbox" error.

Previously, if buildbox-run was present but broken (e.g. segfaulting
on startup), the same fallback codepath would be taken. This lead to
serious breakage in the buildstream-docker-images going undetected:
https://gitlab.com/BuildStream/buildstream-docker-images/-/issues/50

With this change, a broken buildbox-run leads to BuildStream failing
on startup with an error such as this:

    Error instantiating platform: buildbox-run exited with code 5. Output: I am broken!