Update frozen requirements to pick up Python3.11 support in dependencies
The pinned version of some our dependencies (like `pyroaring`) are too
old and hence fail to build on Python 3.11. Here's a sample job that is
failing to install the current list of dependencies:
https://gitlab.com/BuildStream/buildstream-docker-images/-/jobs/3532414193

Note: The reason none of the jobs are failing on this repository is
because none of the current images come with Python 3.11. The `fedora:37`
image would add a test branch for Python 3.11 except that it cannot be
published currently because the tests are failing (like above).