_messenger.py: Use datestampts instead of PID to differentiate log files
Log filenames have traditionally been differentiated by PID in the off chance
that the same cache key is built twice, this however is unreliable, as I can
see in some tests we consistently get the same PID twice in a row (depending
on the system behavior of course).

This commit fixes the cachedfail.py integration test to test correctly,
while the functionality was indeed working correctly, the test was passing
only due to the log file being generated twice in a row with the same
filename.

In order to test this properly, we do the following:

  * Add a new INFO message when discarding a failed build artifact

  * Add a new testing utility Result.get_discarded_elements()