source: disable check for URL marking at configure time
This check is triggered by plugins that have URLs (or URL fragments) as
part of their ref, such as cargo. These plugins currently resort to only
translating part of the URL and appending the rest later.

Ultimately, this check should be investigated and probably only relaxed
rather than dropped entirely.