source: drop check for URL marking at configure time
This check is triggered by plugins that have URLs (or URL fragments)
as part of their ref, such as cargo. These plugins currently resort to
only translating part of the URL and appending the rest
later. Dropping the check allows these plugins to call
mark_download_url() or translate_url() a second time when tracking.

It also allows plugins to postpone calling them until fetch(), which
is the behaviour the check was supposed to prevent. This is still
considered wrong, but will no longer trigger an AssertionError.