source: add suffix parameter to translate_url
Some sources need to download more than one file from a given URL. An
instance of this is the cargo plugin, which needs to download multiple
crates from the same crate registry. The names of the crates come from
the ref, so it can't know all the URLs it needs at configure time (as
they may change with tracking). Instead, it appends the path to the
crate after calling Source.translate_url() with the crate registry URL.
This will be problematic for source mirrors as they need to know the
entire URL to translate it correctly.

This change allows passing the suffix separately so that only the first
part, which may contain an alias, is marked and the entire URL is
translated. This allows keeping in place the check that all URLs are
marked (to check and store the aliases for the purpose of mirroring),
while allowing source plugins to avoid workarounds and thus allowing
the upcoming source mirrors plugins to translate the full URL.