tests/frontend/mirror.py: Test circular dependencies in mirror loading
This tests the behavior of loading a SourceMirror across junction boundaries
when there is a mirror defined for obtaining the source of that same junction.

Tests the successful state of when that mirror is not used, and checks
the failure mode of when the user configuration forces one to use mirrors.