_stream: Only query cache for the requested elements
query_cache() used to query the cache for the elements passed and all
their (build and runtime) dependencies. This is wasteful since most of
the time, we don't need all this cache query.

As it turns out, this is only needed by `bst show` to be able to
figure out the buildable status and the cache key in all circumsances.

This commit makes this behaviour optional, and only enables it from
`bst show`.