_stream: Only query cache for the requested elements
query_cache() used to query the cache for the elements passed and all
their (build and runtime) dependencies. This is wasteful since most of
the time, we don't need all this cache query. And for the cases we
need, it is the job of the query_cache() caller to include the
dependencies they need