node.pyx: Support loading 0/1 as boolean values
It turns out that since OptionBool.get_value() is serializing boolean
values as "0"/"1", we must support this in node parsing.

This is because we should be able to use option value exporting
to values, which can subsequently be specified in plugin configuration
which wants to use MappingNode.get_bool().

Fixes #2006