sandbox: Do not silently drop platform properties
For remote execution, there is currently no way to determine what
platform properties are supported. Silently dropping platform properties
that may or may not be supported can result in build behavior
differences or confusing errors.

This code was originally added before the `build-uid` and `build-gid`
sandbox configuration was made optional to not break remote execution.
However, now that the corresponding platform properties are only set if
requested by the project or element, it's better to fail the build than
to ignore the setting.