Feature: Add support for generating json from bst show command
This adds a new flag called --json to the `bst show` command which when
specified will make bst output a machine readable JSON dump of the dependency
state within the current buildstream project. Having JSON support makes it
easier to write tools that need to inspect the configuration of a buildstream
enabled project.