_sandboxremote.py: stop trying to pull partial artifacts
There used to be a time where buildstream could partially download artifacts,
which could mean that not all artifact blobs are available locally. This is no
longer the case: support for partial artifacts was dropped in favour of the
cache storage-service and is only used for buildtrees, which aren't used in
this case.

Downloading missing blobs from the cache storage-service is handled in
`CASCache._send_blobs()` called a few lines below.