Implement bst inspect subcommand
This adds a new subcommand to bst called `inspect` which dumps structured data
to stdout of a given project. The data sent to stdout can be either JSON or
YAML with JSON being the default.

Having this command available will make writing external tools which need to
inspect the state of a buildstream project more easy since JSON and YAML
encoding are widely supported. This command may easily be extended in the
future to support other inspectable elements related to bst project which
isn't present in this first commit.