tar.py: Fix hardlink extraction with latest Python
Recent Python security fixes (included in 3.13.4 but also backported to
older branches) require link targets to pass the specified filter
function as well. This can result in hardlinked files to not be
extracted when a `base-dir` is set.

This commit separates the filtering from the extraction to avoid this
issue. Pass `filter="tar"` as Python 3.14+ will default to the too
restrictive `filter="data"`.

https://github.com/python/cpython/pull/135037

Fixes #2029.