Still ensure fully loaded if reference is via link
This commit load ensures the toplevel project is always loaded, even
when addressing subproject elements via link.

Commit cff58ec added a call to `Project.ensure_fully_loaded` to ensure
the toplevel project is resolved first when addressing subproject
elements on the command line. The commit assumes that whenever
subproject elements are referenced on the command line,
`Loader.get_loader` is called with `provenance_node == None`.

However, if a link element that points to a subproject element is
referenced on the command line then `get_loader` will be called with
`provenance_node` referencing the link element. This means that
`get_loader` skips the call to `ensure_fully_loaded` causing buildstream
to crash.

This commit fixes the crash by moving the call to `ensure_fully_loaded`
up the call chain into `Loader.load` where we know that the element is
addressed from the command line.