fix: use find_spec for pip plugins
For pip plugins installed in editable mode, via `pip install -e ...`,
buildstream failed to resolve the actual location of source files of a
plugin.
This is caused because `importlib.metadata` does not honor editable
packages with `pth` files.

This patch updates `pluginoriginpip.py` to use `find_spec` to find
files. This works for editable packages.