Support artifact remote arguments in bst artifact show command.
This is useful for ensuring that the shallow artifacts have been downloaded
from the remotes, without requiring heavy duty `bst artifact pull`, and
is particularly useful with the advent of `%{artifact-cas-digest}` being
available in `bst show` output.

This updates the CLI to support the new arguments which we missed
adding previously, and adds the glue code in Stream() object to
load the pipeline with the artifact remote related arguments.