_loader: Merge duplicate dependencies via link elements
Duplicate dependencies are normally merged already in
`extract_depends_from_node()`. However, that only works for dependencies
that are declared with the same filename. If a link element is in a
dependency list, merging may still be required but can be done only
after the link element has been loaded.

The element state updates don't work correctly if there are duplicate
dependencies that aren't merged.

Fixes #2002.